requirements.txt
psycopg2-binary==2.7.4
docopt==0.6.2
